Extract out critical unittest
Additionally, make a fix to the given colorable issue
